Move header out of main-content & nav into header
The old html asssume that the nav element is the 'header' and the
content of the header to be part of the 'main-content' section. More
confusingly, header content has the 'header' tag mark to it.

This change will solve this issue by including the navigation bar & the
header content to all be part of 1 header.