Position the header content to follow the design
Unlike the navigation bar, The header uses 'justify-content: center' to
align its content with gaps added into them.
Attempts at using 'justify-content: space-between' had lead to the
'site-description' and 'banner' to be position at the left & right edge
of the site. This leads to a position that's closer to the design.

Unlike the design however, the navigation logo doesn't exactly align
with the header-content's heading. The current problem is potentially
coming from navigation bar's positioning of its elements rather than the
header-content's.

The navigation bar positioning will be tested by matching
header-content's positining.